requirements.txt
beautifulsoup4
chevron
pyrogram
python-dotenv
telegraph
tgcrypto
